uncrashed doors
the door wasnt the issue, the player would spawn in relation to the camera position and i think it used the position from the last level. to fix this, i just set the players position to the global player x position instead of the camera position.
also i begun adding networking. it isnt quality and i dont even think it works half the time but its networking